This is the first commit on this branch. All of the modifications to the files are done to implement an autotuner.
Added an additional templated function to read Type objects from a
h5file. And show a coresponding example.

Now there is not a whole thermalization sequence when generating a new
Markov Chain. This functionality is "outsourced" to a python autotuner
which calls the "thermalizaiton" step one at a time.

Added thermalFlag parameter to show if a file is thermalized or not.
There is no need for Nmd and Ntherm in yaml file anymore. These two are
determined by the autotuner.